move webview components to a separate file
