+++ src/lib/Components/ArtsyReactWebView.tsx
@@ -314,15 +314,3 @@ export const tracks = {
     context_screen_owner_slug: slug,
   }),
 }
-
-export const RenderTerms: React.FC = () => {
-  return <ArtsyReactWebView url="/terms" />
-}
-
-export const RenderPrivacy: React.FC = () => {
-  return <ArtsyReactWebView url="/privacy" />
-}
-
-export const RenderConditionsOfSale: React.FC = () => {
-  return <ArtsyReactWebView url="/conditions-of-sale" />
-}
